django/requirements.txt
Django==3.1.3
django-allauth==0.43.0
djangorestframework==3.12.1
psycopg2==2.8.6
uwsgi==2.0.19.1
